chore(cmd/riffpad): fold lang_test.go into main_test.go
lang_test.go tested extractLangFlag, which lives in main.go — the orphaned
test filename broke the foo_test.go/foo.go pairing convention and read as if
a lang.go was missing. Merge it into main_test.go, which also gains a real
test instead of holding only the withCliToken helper.
